requirements.txt
dnspython
requests
tqdm